chore: move layout components into layout dir
